Remove config-managed workflow YAML
Before:

  workflows:
    schedules:
      nightly_sync:
        provider: local
        cron: "0 3 * * *"
        target:
          plugin:
            name: roadmap
            operation: sync_items

After:

  POST /api/v1/workflow/schedules
  {
    "provider": "local",
    "cron": "0 3 * * *",
    "target": {
      "plugin": {
        "name": "roadmap",
        "operation": "sync_items"
      }
    }
  }

Delete the top-level workflows config model and startup reconciler so schedules and event triggers are only created through the workflow API or CLI. Agents can still create schedules through workflow system tools. Keep provider-backed schedules, event triggers, events, and runs intact.
