+++ docs/content/providers/authorization.mdx
@@ -46,7 +46,7 @@ so access is both identity-aware and credential-aware.
 Workflow authorization happens twice. Gestalt checks access when a user creates
 a schedule, then checks again when that schedule or trigger fires later. User
 owned workflows keep the creating subject plus a permission ceiling for delayed
-execution. Config-managed workflows use the system config subject instead.
+execution.
 
 ### UI routes
