+++ docs/content/providers/index.mdx
@@ -25,7 +25,7 @@ starts it with the permissions and request context it needs.
 | `runtime` | Hosted execution backends for executable plugins | `runtime.providers.<name>` |
 | `s3` | S3-compatible object stores mounted into plugins | `providers.s3.<name>` |
 | `secrets` | Secret managers that resolve structured secret refs | `providers.secrets.<name>` |
-| `workflow` | Workflow run, schedule, and event-trigger backends | `providers.workflow.<name>` plus top-level `workflows.*` |
+| `workflow` | Workflow run, schedule, and event-trigger backends | `providers.workflow.<name>` |
 | `ui` | Public UI bundles served under configured path prefixes | `providers.ui` |
 
 ## How providers work
