+++ docs/content/reference/built-in-providers.mdx
@@ -119,8 +119,8 @@ plugins:
 ## Workflow Providers
 
 Workflow providers back global runs, schedules, and triggers. They are
-configured under `providers.workflow.<name>`, then referenced by top-level
-`workflows.*` config or the global workflow API/CLI.
+configured under `providers.workflow.<name>`, then selected by the global
+workflow API/CLI.
 
 ```yaml
 providers:
@@ -138,16 +138,6 @@ providers:
         db: workflow
       config:
         pollInterval: 1s
-
-workflows:
-  schedules:
-    nightly_sync:
-      provider: local
-      cron: "0 3 * * *"
-      target:
-        plugin:
-          name: roadmap
-          operation: sync_items
 ```
 
 | Name | Purpose |
