+++ gestaltd/go.mod
@@ -11,7 +11,6 @@ require (
 	github.com/mark3labs/mcp-go v0.49.0
 	github.com/pb33f/libopenapi v0.36.1
 	github.com/prometheus/client_golang v1.23.2
-	github.com/robfig/cron/v3 v3.0.1
 	github.com/santhosh-tekuri/jsonschema/v6 v6.0.2
 	github.com/valon-technologies/gestalt/sdk/go v0.0.0-00010101000000-000000000000
 	go.opentelemetry.io/contrib/bridges/otelslog v0.18.0
