+++ gestaltd/internal/bootstrap/bootstrap.go
@@ -1092,13 +1092,6 @@ func Bootstrap(ctx context.Context, cfg *config.Config, factories *FactoryRegist
 		}
 	}()
 	pluginInvoker.SetTarget(invocation.NewGuarded(sharedInvoker, nil, "plugin", audit, invocation.WithoutRateLimit()))
-	if err := reconcileWorkflowConfigSchedules(ctx, cfg, prepared.Deps.WorkflowRuntime); err != nil {
-		return nil, err
-	}
-	if err := reconcileWorkflowConfigEventTriggers(ctx, cfg, prepared.Deps.WorkflowRuntime); err != nil {
-		return nil, err
-	}
-
 	closeProviders = false
 	closeCore = false
 	closeAudit = false
