+++ gestaltd/internal/bootstrap/bootstrap_test.go
@@ -1752,101 +1752,7 @@ func workflowStartupCallbackConfig(baseURL string) *config.Config {
 	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
 		"temporal": {Source: config.ProviderSource{Path: "stub"}},
 	}
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
 	return cfg
-}
-
-type workflowFixture struct {
-	Provider      string
-	Schedules     map[string]workflowFixtureSchedule
-	EventTriggers map[string]workflowFixtureEventTrigger
-}
-
-type workflowFixtureSchedule struct {
-	Cron      string
-	Timezone  string
-	Operation string
-	Input     map[string]any
-	Paused    bool
-}
-
-type workflowFixtureEventTrigger struct {
-	Match     workflowFixtureEventMatch
-	Operation string
-	Input     map[string]any
-	Paused    bool
-}
-
-type workflowFixtureEventMatch struct {
-	Type    string
-	Source  string
-	Subject string
-}
-
-func setWorkflowFixture(cfg *config.Config, plugin string, workflow *workflowFixture) {
-	if cfg == nil {
-		return
-	}
-	if cfg.Workflows.Schedules == nil {
-		cfg.Workflows.Schedules = map[string]config.WorkflowScheduleConfig{}
-	}
-	if cfg.Workflows.EventTriggers == nil {
-		cfg.Workflows.EventTriggers = map[string]config.WorkflowEventTriggerConfig{}
-	}
-	for key, schedule := range cfg.Workflows.Schedules {
-		if workflowFixtureTargetPlugin(schedule.Target) == plugin {
-			delete(cfg.Workflows.Schedules, key)
-		}
-	}
-	for key, trigger := range cfg.Workflows.EventTriggers {
-		if workflowFixtureTargetPlugin(trigger.Target) == plugin {
-			delete(cfg.Workflows.EventTriggers, key)
-		}
-	}
-	if workflow == nil {
-		return
-	}
-	for key, schedule := range workflow.Schedules {
-		cfg.Workflows.Schedules[key] = config.WorkflowScheduleConfig{
-			Provider: workflow.Provider,
-			Target:   workflowFixtureTarget(plugin, schedule.Operation, schedule.Input),
-			Cron:     schedule.Cron,
-			Timezone: schedule.Timezone,
-			Paused:   schedule.Paused,
-		}
-	}
-	for key, trigger := range workflow.EventTriggers {
-		cfg.Workflows.EventTriggers[key] = config.WorkflowEventTriggerConfig{
-			Provider: workflow.Provider,
-			Target:   workflowFixtureTarget(plugin, trigger.Operation, trigger.Input),
-			Match: config.WorkflowEventMatch{
-				Type:    trigger.Match.Type,
-				Source:  trigger.Match.Source,
-				Subject: trigger.Match.Subject,
-			},
-			Paused: trigger.Paused,
-		}
-	}
-}
-
-func workflowFixtureTarget(plugin, operation string, input map[string]any) *config.WorkflowTargetConfig {
-	return &config.WorkflowTargetConfig{
-		Plugin: &config.WorkflowPluginTargetConfig{
-			Name:      plugin,
-			Operation: operation,
-			Input:     maps.Clone(input),
-		},
-	}
-}
-
-func requireCoreWorkflowPluginTarget(t *testing.T, target coreworkflow.Target) *coreworkflow.PluginTarget {
-	t.Helper()
-	if target.Plugin == nil {
-		t.Fatalf("target plugin is nil: %#v", target)
-	}
-	return target.Plugin
 }
 
 func coreWorkflowPluginTarget(pluginName, operation string) coreworkflow.Target {
@@ -1867,13 +1773,6 @@ func protoWorkflowPluginTarget(pluginName, operation string) *proto.BoundWorkflo
 			},
 		},
 	}
-}
-
-func workflowFixtureTargetPlugin(target *config.WorkflowTargetConfig) string {
-	if target == nil || target.Plugin == nil {
-		return ""
-	}
-	return target.Plugin.Name
 }
 
 func transportSecretRef(name string) string {
@@ -4253,1482 +4152,6 @@ func TestBootstrapRoutesWorkflowIndexedDBHostServices(t *testing.T) {
 	}
 }
 
-func TestBootstrapAppliesConfiguredWorkflowSchedules(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "America/New_York",
-				Operation: "sync",
-				Input: map[string]any{
-					"source": "yaml",
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	factories := validFactories()
-	recorders := map[string]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{}
-		recorders[name] = recorder
-		return recorder, nil
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	recorder := recorders["temporal"]
-	if recorder == nil {
-		t.Fatal("missing workflow recorder for temporal")
-		return
-	}
-	if len(recorder.upsertedSchedules) != 1 {
-		t.Fatalf("upserted schedules = %d, want 1", len(recorder.upsertedSchedules))
-	}
-	got := recorder.upsertedSchedules[0]
-	if got.ScheduleID != workflowConfigScheduleID("nightly_sync") {
-		t.Fatalf("schedule id = %q", got.ScheduleID)
-	}
-	if got.Cron != "0 2 * * *" || got.Timezone != "America/New_York" {
-		t.Fatalf("schedule timing = %#v", got)
-	}
-	gotPlugin := requireCoreWorkflowPluginTarget(t, got.Target)
-	if gotPlugin.PluginName != "roadmap" || gotPlugin.Operation != "sync" {
-		t.Fatalf("target = %#v", got.Target)
-	}
-	if gotPlugin.Input["source"] != "yaml" {
-		t.Fatalf("target input = %#v", gotPlugin.Input)
-	}
-	if got.RequestedBy.SubjectID != "system:config" || got.RequestedBy.SubjectKind != "system" || got.RequestedBy.AuthSource != "config" {
-		t.Fatalf("requestedBy = %#v", got.RequestedBy)
-	}
-	if strings.TrimSpace(got.ExecutionRef) == "" {
-		t.Fatal("execution ref = empty")
-	}
-	ref, err := recorder.GetExecutionReference(context.Background(), got.ExecutionRef)
-	if err != nil {
-		t.Fatalf("Get execution ref: %v", err)
-	}
-	if ref.SubjectID != "system:config" {
-		t.Fatalf("subjectID = %q, want %q", ref.SubjectID, "system:config")
-	}
-	if len(ref.Permissions) != 1 || ref.Permissions[0].Plugin != "roadmap" || len(ref.Permissions[0].Operations) != 1 || ref.Permissions[0].Operations[0] != "sync" {
-		t.Fatalf("permissions = %#v", ref.Permissions)
-	}
-}
-
-func TestValidateDoesNotApplyConfiguredWorkflowSchedules(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Paused:    true,
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	factories := validFactories()
-	recorders := map[string]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{}
-		recorders[name] = recorder
-		return recorder, nil
-	}
-
-	if _, err := bootstrap.Validate(context.Background(), cfg, factories); err != nil {
-		t.Fatalf("Validate: %v", err)
-	}
-
-	recorder := recorders["temporal"]
-	if recorder == nil {
-		t.Fatal("missing workflow recorder for temporal")
-		return
-	}
-	if len(recorder.upsertedSchedules) != 0 {
-		t.Fatalf("upserted schedules = %d, want 0", len(recorder.upsertedSchedules))
-	}
-	if len(recorder.deletedSchedules) != 0 {
-		t.Fatalf("deleted schedules = %d, want 0", len(recorder.deletedSchedules))
-	}
-}
-
-func TestBootstrapRejectsConfiguredWorkflowSchedulesForUserCredentialedPlugins(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	cfg.Plugins["roadmap"].ConnectionMode = providermanifestv1.ConnectionModeUser
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-
-	factories := validFactories()
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return &recordingWorkflowProvider{}, nil
-	}
-
-	_, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err == nil {
-		t.Fatal("expected Bootstrap to reject user-credentialed config-managed schedules")
-	}
-	if !strings.Contains(err.Error(), `config-managed workflows do not support user-credentialed plugin "roadmap"`) {
-		t.Fatalf("Bootstrap error = %v", err)
-	}
-}
-
-func TestBootstrapDeletesRemovedConfiguredWorkflowSchedules(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	recorders := []*recordingWorkflowProvider{}
-	sharedSchedules := map[string]*coreworkflow.Schedule{}
-	sharedExecutionRefs := map[string]*coreworkflow.ExecutionReference{}
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{
-			schedules:     sharedSchedules,
-			executionRefs: sharedExecutionRefs,
-		}
-		recorders = append(recorders, recorder)
-		return recorder, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Input: map[string]any{
-					"limit": 1,
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	if len(recorders) != 1 || len(recorders[0].upsertedSchedules) != 1 {
-		t.Fatalf("initial upserts = %#v", recorders)
-	}
-	initialExecutionRef := recorders[0].upsertedSchedules[0].ExecutionRef
-	_ = result.Close(context.Background())
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove schedule: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(recorders) != 2 {
-		t.Fatalf("recorders = %d, want 2", len(recorders))
-	}
-	staleID := workflowConfigScheduleID("nightly_sync")
-	recorder := recorders[1]
-	if len(recorder.deletedSchedules) != 1 {
-		t.Fatalf("deleted schedules = %d, want 1", len(recorder.deletedSchedules))
-	}
-	if recorder.deletedSchedules[0].ScheduleID != staleID {
-		t.Fatalf("delete request = %#v", recorder.deletedSchedules[0])
-	}
-	if len(recorder.upsertedSchedules) != 0 {
-		t.Fatalf("upserted schedules = %d, want 0", len(recorder.upsertedSchedules))
-	}
-	ref, err := recorder.GetExecutionReference(context.Background(), initialExecutionRef)
-	if err != nil {
-		t.Fatalf("Get revoked execution ref: %v", err)
-	}
-	if ref.RevokedAt == nil || ref.RevokedAt.IsZero() {
-		t.Fatalf("revokedAt = %#v, want set", ref.RevokedAt)
-	}
-}
-
-func TestBootstrapIgnoresUserSchedulesThatOnlyShareCfgPrefix(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	factories := validFactories()
-	recorders := map[string]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{
-			listedSchedules: []*coreworkflow.Schedule{{ID: "cfg_backup"}},
-		}
-		recorders[name] = recorder
-		return recorder, nil
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	recorder := recorders["temporal"]
-	if recorder == nil {
-		t.Fatal("missing workflow recorder for temporal")
-		return
-	}
-	if len(recorder.deletedSchedules) != 0 {
-		t.Fatalf("deleted schedules = %d, want 0", len(recorder.deletedSchedules))
-	}
-}
-
-func TestBootstrapMovesConfiguredWorkflowSchedulesToNewProvider(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	recorders := map[string][]*recordingWorkflowProvider{}
-	sharedSchedules := map[string]map[string]*coreworkflow.Schedule{}
-	sharedExecutionRefs := map[string]map[string]*coreworkflow.ExecutionReference{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		if sharedSchedules[name] == nil {
-			sharedSchedules[name] = map[string]*coreworkflow.Schedule{}
-		}
-		if sharedExecutionRefs[name] == nil {
-			sharedExecutionRefs[name] = map[string]*coreworkflow.ExecutionReference{}
-		}
-		recorder := &recordingWorkflowProvider{
-			schedules:     sharedSchedules[name],
-			executionRefs: sharedExecutionRefs[name],
-		}
-		recorders[name] = append(recorders[name], recorder)
-		return recorder, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Input: map[string]any{
-					"limit": 1,
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	if len(recorders["temporal"]) != 1 || len(recorders["temporal"][0].upsertedSchedules) != 1 {
-		t.Fatalf("initial temporal recorders = %#v", recorders["temporal"])
-	}
-	initialExecutionRef := recorders["temporal"][0].upsertedSchedules[0].ExecutionRef
-	_ = result.Close(context.Background())
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "backup",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Input: map[string]any{
-					"limit": 1,
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap move provider: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(recorders["temporal"]) != 2 || len(recorders["backup"]) != 2 {
-		t.Fatalf("recorders = %#v", recorders)
-	}
-	if len(recorders["temporal"][1].deletedSchedules) != 1 {
-		t.Fatalf("temporal deleted schedules = %d, want 1", len(recorders["temporal"][1].deletedSchedules))
-	}
-	if len(recorders["backup"][1].upsertedSchedules) != 1 {
-		t.Fatalf("backup upserted schedules = %d, want 1", len(recorders["backup"][1].upsertedSchedules))
-	}
-	backupExecutionRef := recorders["backup"][1].upsertedSchedules[0].ExecutionRef
-	oldRef, err := recorders["temporal"][1].GetExecutionReference(context.Background(), initialExecutionRef)
-	if err != nil {
-		t.Fatalf("Get initial execution ref: %v", err)
-	}
-	if oldRef.RevokedAt == nil || oldRef.RevokedAt.IsZero() {
-		t.Fatalf("initial revokedAt = %#v, want set", oldRef.RevokedAt)
-	}
-	newRef, err := recorders["backup"][1].GetExecutionReference(context.Background(), backupExecutionRef)
-	if err != nil {
-		t.Fatalf("Get backup execution ref: %v", err)
-	}
-	if newRef.ProviderName != "backup" {
-		t.Fatalf("providerName = %q, want %q", newRef.ProviderName, "backup")
-	}
-	if newRef.RevokedAt != nil {
-		t.Fatalf("backup revokedAt = %#v, want nil", newRef.RevokedAt)
-	}
-}
-
-func TestBootstrapClosesWorkflowProvidersWhenConfigScheduleReconcileFails(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	closed := &atomic.Bool{}
-	db := &coretesting.StubIndexedDB{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	temporalSchedules := map[string]*coreworkflow.Schedule{}
-	temporalExecutionRefs := map[string]*coreworkflow.ExecutionReference{}
-	temporalStarts := 0
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		if name == "temporal" {
-			temporalStarts++
-			provider := &recordingWorkflowProvider{
-				schedules:     temporalSchedules,
-				executionRefs: temporalExecutionRefs,
-				closed:        closed,
-			}
-			if temporalStarts > 1 {
-				provider.deleteScheduleErr = fmt.Errorf("delete boom")
-			}
-			return provider, nil
-		}
-		return &recordingWorkflowProvider{closed: closed}, nil
-	}
-
-	cfg.Workflows.Schedules = map[string]config.WorkflowScheduleConfig{
-		"nightly_sync": {
-			Target:   workflowFixtureTarget("roadmap", "sync", nil),
-			Cron:     "0 2 * * *",
-			Timezone: "UTC",
-		},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	_ = result.Close(context.Background())
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "backup",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Input: map[string]any{
-					"limit": 1,
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	_, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err == nil || !strings.Contains(err.Error(), "delete boom") {
-		t.Fatalf("Bootstrap error = %v, want delete failure", err)
-	}
-	if !closed.Load() {
-		t.Fatal("workflow provider was not closed after reconcile failure")
-	}
-}
-
-func TestBootstrapDoesNotApplyConfiguredWorkflowSchedulesWhenAuditBuildFails(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Input: map[string]any{
-					"limit": 1,
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-	cfg.Providers.Audit = map[string]*config.ProviderEntry{
-		"default": {Source: config.ProviderSource{Builtin: "test-audit"}},
-	}
-
-	factories := validFactories()
-	recorder := &recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return recorder, nil
-	}
-	factories.Audit = func(context.Context, config.ProviderEntry, core.TelemetryProvider) (core.AuditSink, func(context.Context) error, error) {
-		return nil, nil, fmt.Errorf("audit boom")
-	}
-
-	_, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err == nil || !strings.Contains(err.Error(), "audit boom") {
-		t.Fatalf("Bootstrap error = %v, want audit failure", err)
-	}
-	if len(recorder.upsertedSchedules) != 0 {
-		t.Fatalf("upserted schedules = %d, want 0", len(recorder.upsertedSchedules))
-	}
-}
-
-func TestBootstrapRejectsExistingUnmanagedWorkflowScheduleID(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	recorder := &recordingWorkflowProvider{
-		getSchedule: &coreworkflow.Schedule{
-			ID:       workflowConfigScheduleID("nightly_sync"),
-			Cron:     "0 2 * * *",
-			Timezone: "UTC",
-			Target:   coreWorkflowPluginTarget("roadmap", "sync"),
-		},
-	}
-	factories := validFactories()
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return recorder, nil
-	}
-
-	_, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err == nil || !strings.Contains(err.Error(), "conflicts with existing unmanaged schedule id") {
-		t.Fatalf("Bootstrap error = %v, want ownership conflict", err)
-	}
-	if len(recorder.upsertedSchedules) != 0 {
-		t.Fatalf("upserted schedules = %d, want 0", len(recorder.upsertedSchedules))
-	}
-}
-
-func TestBootstrapReusesConfiguredWorkflowExecutionRefAcrossUnchangedBootstrap(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	provider := &recordingWorkflowProvider{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return provider, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-				Input: map[string]any{
-					"limit": 1,
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	initialExecutionRef := provider.upsertedSchedules[0].ExecutionRef
-	provider.executionRefs[initialExecutionRef].Target.Plugin.Input["limit"] = float64(1)
-	_ = result.Close(context.Background())
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap replay: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(provider.upsertedSchedules) != 2 {
-		t.Fatalf("upserted schedules = %d, want 2", len(provider.upsertedSchedules))
-	}
-	reusedExecutionRef := provider.upsertedSchedules[1].ExecutionRef
-	if reusedExecutionRef != initialExecutionRef {
-		t.Fatalf("execution ref = %q, want reuse of %q", reusedExecutionRef, initialExecutionRef)
-	}
-	ref, err := provider.GetExecutionReference(context.Background(), initialExecutionRef)
-	if err != nil {
-		t.Fatalf("Get execution ref: %v", err)
-	}
-	if ref.RevokedAt != nil {
-		t.Fatalf("revokedAt = %#v, want nil", ref.RevokedAt)
-	}
-}
-
-func TestBootstrapRefreshesConfiguredWorkflowExecutionRefWhenMetadataChanges(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	provider := &recordingWorkflowProvider{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return provider, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	initialExecutionRef := provider.upsertedSchedules[0].ExecutionRef
-	_ = result.Close(context.Background())
-	provider.executionRefs[initialExecutionRef].DisplayName = "Stale config name"
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap replay: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(provider.upsertedSchedules) != 2 {
-		t.Fatalf("upserted schedules = %d, want 2", len(provider.upsertedSchedules))
-	}
-	refreshedExecutionRef := provider.upsertedSchedules[1].ExecutionRef
-	if refreshedExecutionRef == initialExecutionRef {
-		t.Fatalf("execution ref = %q, want refreshed ref after display name drift", refreshedExecutionRef)
-	}
-	ref, err := provider.GetExecutionReference(context.Background(), refreshedExecutionRef)
-	if err != nil {
-		t.Fatalf("Get refreshed execution ref: %v", err)
-	}
-	if ref.DisplayName != "Gestalt config" {
-		t.Fatalf("displayName = %q, want refreshed config display name", ref.DisplayName)
-	}
-	oldRef, err := provider.GetExecutionReference(context.Background(), initialExecutionRef)
-	if err != nil {
-		t.Fatalf("Get old execution ref: %v", err)
-	}
-	if oldRef.RevokedAt == nil || oldRef.RevokedAt.IsZero() {
-		t.Fatalf("old revokedAt = %#v, want set", oldRef.RevokedAt)
-	}
-}
-
-func TestBootstrapIgnoresMissingRemovedConfiguredWorkflowSchedule(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	provider := &recordingWorkflowProvider{deleteMissingNotFound: true}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return provider, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	_ = result.Close(context.Background())
-	provider.schedules = map[string]*coreworkflow.Schedule{}
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove missing schedule: %v", err)
-	}
-	_ = result.Close(context.Background())
-
-	if len(provider.deletedSchedules) != 0 {
-		t.Fatalf("deleted schedules = %d, want 0", len(provider.deletedSchedules))
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove missing schedule replay: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(provider.deletedSchedules) != 0 {
-		t.Fatalf("deleted schedules after replay = %d, want 0", len(provider.deletedSchedules))
-	}
-}
-
-func TestBootstrapIgnoresMissingPreviousScheduleDuringWorkflowProviderMove(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	temporal := &recordingWorkflowProvider{deleteMissingNotFound: true}
-	backup := &recordingWorkflowProvider{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		if name == "backup" {
-			return backup, nil
-		}
-		return temporal, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	_ = result.Close(context.Background())
-	temporal.schedules = map[string]*coreworkflow.Schedule{}
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "backup",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap move provider: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(backup.upsertedSchedules) != 1 {
-		t.Fatalf("backup upserted schedules = %d, want 1", len(backup.upsertedSchedules))
-	}
-	if len(temporal.deletedSchedules) != 0 {
-		t.Fatalf("temporal deleted schedules = %d, want 0", len(temporal.deletedSchedules))
-	}
-}
-
-func TestBootstrapDeletesRemovedConfiguredWorkflowSchedulesWhenProviderDropsExecutionRef(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	provider := &recordingWorkflowProvider{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return provider, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Timezone:  "UTC",
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	initialExecutionRef := provider.upsertedSchedules[0].ExecutionRef
-	_ = result.Close(context.Background())
-
-	provider.omitScheduleExecutionRef = true
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove schedule: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	ref, err := provider.GetExecutionReference(context.Background(), initialExecutionRef)
-	if err != nil {
-		t.Fatalf("Get revoked execution ref: %v", err)
-	}
-	if ref.RevokedAt != nil {
-		t.Fatalf("revokedAt = %#v, want nil when provider omits execution_ref", ref.RevokedAt)
-	}
-}
-
-func TestBootstrapAppliesConfiguredWorkflowEventTriggers(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type:   "task.updated",
-					Source: "roadmap",
-				},
-				Operation: "sync",
-				Input: map[string]any{
-					"source": "yaml",
-				},
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	factories := validFactories()
-	recorders := map[string]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{}
-		recorders[name] = recorder
-		return recorder, nil
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	recorder := recorders["temporal"]
-	if recorder == nil {
-		t.Fatal("missing workflow recorder for temporal")
-		return
-	}
-	if len(recorder.upsertedEventTriggers) != 1 {
-		t.Fatalf("upserted event triggers = %d, want 1", len(recorder.upsertedEventTriggers))
-	}
-	got := recorder.upsertedEventTriggers[0]
-	if got.TriggerID != workflowConfigEventTriggerID("task_updated") {
-		t.Fatalf("trigger id = %q", got.TriggerID)
-	}
-	if got.Match.Type != "task.updated" || got.Match.Source != "roadmap" || got.Match.Subject != "" {
-		t.Fatalf("match = %#v", got.Match)
-	}
-	gotPlugin := requireCoreWorkflowPluginTarget(t, got.Target)
-	if gotPlugin.PluginName != "roadmap" || gotPlugin.Operation != "sync" {
-		t.Fatalf("target = %#v", got.Target)
-	}
-	if gotPlugin.Input["source"] != "yaml" {
-		t.Fatalf("target input = %#v", gotPlugin.Input)
-	}
-	if got.RequestedBy.SubjectID != "system:config" || got.RequestedBy.SubjectKind != "system" || got.RequestedBy.AuthSource != "config" {
-		t.Fatalf("requestedBy = %#v", got.RequestedBy)
-	}
-	if strings.TrimSpace(got.ExecutionRef) == "" {
-		t.Fatal("execution ref = empty")
-	}
-	ref, err := recorder.GetExecutionReference(context.Background(), got.ExecutionRef)
-	if err != nil {
-		t.Fatalf("Get execution ref: %v", err)
-	}
-	if ref.SubjectID != "system:config" {
-		t.Fatalf("subjectID = %q, want %q", ref.SubjectID, "system:config")
-	}
-}
-
-func TestBootstrapConfigManagedAgentTargetsPreserveWorkflowSystemToolRefs(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	cfg.Providers.Agent = map[string]*config.ProviderEntry{
-		"managed": {Source: config.ProviderSource{Path: "stub"}},
-	}
-	agentTarget := &config.WorkflowTargetConfig{Agent: &config.WorkflowAgentConfig{
-		Provider: "managed",
-		Prompt:   "Inspect the workflow and sync the roadmap",
-		Tools: []config.WorkflowAgentToolRef{
-			{System: coreagent.SystemToolWorkflow, Operation: "schedules.list"},
-			{Plugin: "roadmap", Operation: "sync"},
-		},
-	}}
-	cfg.Workflows.Schedules = map[string]config.WorkflowScheduleConfig{
-		"agent_schedule": {
-			Provider: "temporal",
-			Cron:     "*/10 * * * *",
-			Timezone: "UTC",
-			Target:   agentTarget,
-		},
-	}
-	cfg.Workflows.EventTriggers = map[string]config.WorkflowEventTriggerConfig{
-		"agent_event": {
-			Provider: "temporal",
-			Match: config.WorkflowEventMatch{
-				Type: "roadmap.updated",
-			},
-			Target: agentTarget,
-		},
-	}
-
-	factories := validFactories()
-	factories.Agent = func(context.Context, string, yaml.Node, []providerhost.HostService, bootstrap.Deps) (coreagent.Provider, error) {
-		return newRecordingAgentProvider(), nil
-	}
-	recorders := map[string]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{}
-		recorders[name] = recorder
-		return recorder, nil
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	recorder := recorders["temporal"]
-	if recorder == nil {
-		t.Fatal("missing workflow recorder for temporal")
-		return
-	}
-	if len(recorder.upsertedSchedules) != 1 {
-		t.Fatalf("upserted schedules = %d, want 1", len(recorder.upsertedSchedules))
-	}
-	if len(recorder.upsertedEventTriggers) != 1 {
-		t.Fatalf("upserted event triggers = %d, want 1", len(recorder.upsertedEventTriggers))
-	}
-	for label, target := range map[string]coreworkflow.Target{
-		"schedule":      recorder.upsertedSchedules[0].Target,
-		"event trigger": recorder.upsertedEventTriggers[0].Target,
-	} {
-		if target.Agent == nil || len(target.Agent.ToolRefs) != 2 {
-			t.Fatalf("%s target = %#v", label, target)
-		}
-		if target.Agent.ToolRefs[0].System != coreagent.SystemToolWorkflow || target.Agent.ToolRefs[0].Operation != "schedules.list" {
-			t.Fatalf("%s workflow tool ref = %#v", label, target.Agent.ToolRefs[0])
-		}
-		if target.Agent.ToolRefs[1].Plugin != "roadmap" || target.Agent.ToolRefs[1].Operation != "sync" {
-			t.Fatalf("%s plugin tool ref = %#v", label, target.Agent.ToolRefs[1])
-		}
-	}
-	for _, executionRef := range []string{
-		recorder.upsertedSchedules[0].ExecutionRef,
-		recorder.upsertedEventTriggers[0].ExecutionRef,
-	} {
-		ref, err := recorder.GetExecutionReference(context.Background(), executionRef)
-		if err != nil {
-			t.Fatalf("Get execution ref %q: %v", executionRef, err)
-		}
-		if len(ref.Permissions) != 1 || ref.Permissions[0].Plugin != "roadmap" || len(ref.Permissions[0].Operations) != 1 || ref.Permissions[0].Operations[0] != "sync" {
-			t.Fatalf("permissions = %#v", ref.Permissions)
-		}
-	}
-}
-
-func TestValidateDoesNotApplyConfiguredWorkflowEventTriggers(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-				Paused:    true,
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	factories := validFactories()
-	recorders := map[string]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{}
-		recorders[name] = recorder
-		return recorder, nil
-	}
-
-	if _, err := bootstrap.Validate(context.Background(), cfg, factories); err != nil {
-		t.Fatalf("Validate: %v", err)
-	}
-
-	recorder := recorders["temporal"]
-	if recorder == nil {
-		t.Fatal("missing workflow recorder for temporal")
-		return
-	}
-	if len(recorder.upsertedEventTriggers) != 0 {
-		t.Fatalf("upserted event triggers = %d, want 0", len(recorder.upsertedEventTriggers))
-	}
-	if len(recorder.deletedEventTriggers) != 0 {
-		t.Fatalf("deleted event triggers = %d, want 0", len(recorder.deletedEventTriggers))
-	}
-}
-
-func TestBootstrapDeletesRemovedConfiguredWorkflowEventTriggers(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	recorders := []*recordingWorkflowProvider{}
-	sharedEventTriggers := map[string]*coreworkflow.EventTrigger{}
-	sharedExecutionRefs := map[string]*coreworkflow.ExecutionReference{}
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{
-			eventTriggers: sharedEventTriggers,
-			executionRefs: sharedExecutionRefs,
-		}
-		recorders = append(recorders, recorder)
-		return recorder, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	if len(recorders) != 1 || len(recorders[0].upsertedEventTriggers) != 1 {
-		t.Fatalf("initial upserts = %#v", recorders)
-	}
-	_ = result.Close(context.Background())
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove event trigger: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(recorders) != 2 {
-		t.Fatalf("recorders = %d, want 2", len(recorders))
-	}
-	staleID := workflowConfigEventTriggerID("task_updated")
-	recorder := recorders[1]
-	if len(recorder.deletedEventTriggers) != 1 {
-		t.Fatalf("deleted event triggers = %d, want 1", len(recorder.deletedEventTriggers))
-	}
-	if recorder.deletedEventTriggers[0].TriggerID != staleID {
-		t.Fatalf("delete request = %#v", recorder.deletedEventTriggers[0])
-	}
-	if len(recorder.upsertedEventTriggers) != 0 {
-		t.Fatalf("upserted event triggers = %d, want 0", len(recorder.upsertedEventTriggers))
-	}
-}
-
-func TestBootstrapMovesConfiguredWorkflowEventTriggersToNewProvider(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	recorders := map[string][]*recordingWorkflowProvider{}
-	sharedEventTriggers := map[string]map[string]*coreworkflow.EventTrigger{}
-	sharedExecutionRefs := map[string]map[string]*coreworkflow.ExecutionReference{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		if sharedEventTriggers[name] == nil {
-			sharedEventTriggers[name] = map[string]*coreworkflow.EventTrigger{}
-		}
-		if sharedExecutionRefs[name] == nil {
-			sharedExecutionRefs[name] = map[string]*coreworkflow.ExecutionReference{}
-		}
-		recorder := &recordingWorkflowProvider{
-			eventTriggers: sharedEventTriggers[name],
-			executionRefs: sharedExecutionRefs[name],
-		}
-		recorders[name] = append(recorders[name], recorder)
-		return recorder, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	if len(recorders["temporal"]) != 1 || len(recorders["temporal"][0].upsertedEventTriggers) != 1 {
-		t.Fatalf("initial temporal recorders = %#v", recorders["temporal"])
-	}
-	_ = result.Close(context.Background())
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "backup",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap move provider: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(recorders["temporal"]) != 2 || len(recorders["backup"]) != 2 {
-		t.Fatalf("recorders = %#v", recorders)
-	}
-	if len(recorders["temporal"][1].deletedEventTriggers) != 1 {
-		t.Fatalf("temporal deleted event triggers = %d, want 1", len(recorders["temporal"][1].deletedEventTriggers))
-	}
-	if len(recorders["backup"][1].upsertedEventTriggers) != 1 {
-		t.Fatalf("backup upserted event triggers = %d, want 1", len(recorders["backup"][1].upsertedEventTriggers))
-	}
-}
-
-func TestBootstrapRejectsExistingUnmanagedWorkflowEventTriggerIDDuringProviderMove(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	recorders := map[string][]*recordingWorkflowProvider{}
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		recorder := &recordingWorkflowProvider{}
-		if name == "backup" && len(recorders[name]) == 1 {
-			recorder.getEventTrigger = &coreworkflow.EventTrigger{ID: workflowConfigEventTriggerID("task_updated")}
-		}
-		recorders[name] = append(recorders[name], recorder)
-		return recorder, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	_ = result.Close(context.Background())
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "backup",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	_, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err == nil || !strings.Contains(err.Error(), "conflicts with existing unmanaged trigger id") {
-		t.Fatalf("Bootstrap error = %v, want ownership conflict", err)
-	}
-	if len(recorders["backup"]) != 2 {
-		t.Fatalf("backup recorders = %d, want 2", len(recorders["backup"]))
-	}
-	if len(recorders["backup"][1].upsertedEventTriggers) != 0 {
-		t.Fatalf("backup upserted event triggers = %d, want 0", len(recorders["backup"][1].upsertedEventTriggers))
-	}
-}
-
-func TestBootstrapRejectsExistingUnmanagedWorkflowEventTriggerID(t *testing.T) {
-	t.Parallel()
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	recorder := &recordingWorkflowProvider{
-		getEventTrigger: &coreworkflow.EventTrigger{
-			ID: workflowConfigEventTriggerID("task_updated"),
-			Match: coreworkflow.EventMatch{
-				Type: "task.updated",
-			},
-			Target: coreWorkflowPluginTarget("roadmap", "sync"),
-		},
-	}
-	factories := validFactories()
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return recorder, nil
-	}
-
-	_, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err == nil || !strings.Contains(err.Error(), "conflicts with existing unmanaged trigger id") {
-		t.Fatalf("Bootstrap error = %v, want ownership conflict", err)
-	}
-	if len(recorder.upsertedEventTriggers) != 0 {
-		t.Fatalf("upserted event triggers = %d, want 0", len(recorder.upsertedEventTriggers))
-	}
-}
-
-func TestBootstrapIgnoresMissingRemovedConfiguredWorkflowEventTrigger(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	provider := &recordingWorkflowProvider{deleteEventMissingNotFound: true}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, _ string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		return provider, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	_ = result.Close(context.Background())
-	provider.eventTriggers = map[string]*coreworkflow.EventTrigger{}
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove missing event trigger: %v", err)
-	}
-	_ = result.Close(context.Background())
-
-	if len(provider.deletedEventTriggers) != 0 {
-		t.Fatalf("deleted event triggers = %d, want 0", len(provider.deletedEventTriggers))
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap remove missing event trigger replay: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(provider.deletedEventTriggers) != 0 {
-		t.Fatalf("deleted event triggers after replay = %d, want 0", len(provider.deletedEventTriggers))
-	}
-}
-
-func TestBootstrapIgnoresMissingPreviousEventTriggerDuringWorkflowProviderMove(t *testing.T) {
-	t.Parallel()
-
-	db := &coretesting.StubIndexedDB{}
-	temporal := &recordingWorkflowProvider{deleteEventMissingNotFound: true}
-	backup := &recordingWorkflowProvider{}
-	factories := validFactories()
-	factories.IndexedDB = func(yaml.Node) (indexeddb.IndexedDB, error) { return db, nil }
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		if name == "backup" {
-			return backup, nil
-		}
-		return temporal, nil
-	}
-
-	cfg := workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap initial: %v", err)
-	}
-	_ = result.Close(context.Background())
-	temporal.eventTriggers = map[string]*coreworkflow.EventTrigger{}
-
-	cfg = workflowStartupCallbackConfig("https://example.invalid")
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "backup",
-		EventTriggers: map[string]workflowFixtureEventTrigger{
-			"task_updated": {
-				Match: workflowFixtureEventMatch{
-					Type: "task.updated",
-				},
-				Operation: "sync",
-			},
-		},
-	})
-	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
-		"temporal": {Source: config.ProviderSource{Path: "stub"}},
-		"backup":   {Source: config.ProviderSource{Path: "stub"}},
-	}
-
-	result, err = bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap move provider: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(backup.upsertedEventTriggers) != 1 {
-		t.Fatalf("backup upserted event triggers = %d, want 1", len(backup.upsertedEventTriggers))
-	}
-	if len(temporal.deletedEventTriggers) != 0 {
-		t.Fatalf("temporal deleted event triggers = %d, want 0", len(temporal.deletedEventTriggers))
-	}
-}
-
-func workflowConfigScheduleID(scheduleKey string) string {
-	sum := sha256.Sum256([]byte(scheduleKey))
-	return coreworkflow.ConfigManagedSchedulePrefix + hex.EncodeToString(sum[:])
-}
-
-func workflowConfigEventTriggerID(triggerKey string) string {
-	sum := sha256.Sum256([]byte("event_trigger\x00" + triggerKey))
-	return coreworkflow.ConfigManagedSchedulePrefix + hex.EncodeToString(sum[:])
-}
-
 func TestBootstrapStartsWorkflowProvidersAfterInvokerIsReady(t *testing.T) {
 	t.Parallel()
 
@@ -6324,79 +4747,6 @@ func TestBootstrapAgentProviderRejectsMismatchedRequestedSessionOrTurnID(t *test
 	}
 }
 
-func TestBootstrapConfiguredWorkflowScheduleExecutionRefInvokesPolicyProtectedPlugin(t *testing.T) {
-	t.Parallel()
-
-	var requestPath atomic.Value
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		requestPath.Store(r.URL.Path)
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusAccepted)
-		_, _ = w.Write([]byte(`{"ok":true}`))
-	}))
-	defer srv.Close()
-
-	cfg := workflowStartupCallbackConfig(srv.URL)
-	cfg.Plugins["roadmap"].AuthorizationPolicy = "roadmap-policy"
-	cfg.Plugins["roadmap"].ResolvedManifest.Spec.Surfaces.REST.Operations[0].AllowedRoles = []string{"viewer"}
-	cfg.Authorization.Policies = map[string]config.SubjectPolicyDef{
-		"roadmap-policy": {
-			Members: []config.SubjectPolicyMemberDef{{
-				SubjectID: principal.UserSubjectID("viewer-user"),
-				Role:      "viewer",
-			}},
-		},
-	}
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-		Schedules: map[string]workflowFixtureSchedule{
-			"nightly_sync": {
-				Cron:      "0 2 * * *",
-				Operation: "sync",
-			},
-		},
-	})
-
-	recorder := &recordingWorkflowProvider{}
-	var hostServices []providerhost.HostService
-	factories := validFactories()
-	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, services []providerhost.HostService, _ bootstrap.Deps) (coreworkflow.Provider, error) {
-		if name != "temporal" {
-			return nil, fmt.Errorf("workflow name = %q, want %q", name, "temporal")
-		}
-		hostServices = append([]providerhost.HostService(nil), services...)
-		return recorder, nil
-	}
-
-	result, err := bootstrap.Bootstrap(context.Background(), cfg, factories)
-	if err != nil {
-		t.Fatalf("Bootstrap: %v", err)
-	}
-	defer func() { _ = result.Close(context.Background()) }()
-	<-result.ProvidersReady
-
-	if len(recorder.upsertedSchedules) != 1 {
-		t.Fatalf("upserted schedules = %d, want 1", len(recorder.upsertedSchedules))
-	}
-	executionRef := recorder.upsertedSchedules[0].ExecutionRef
-	if executionRef == "" {
-		t.Fatal("execution ref is empty")
-	}
-	resp, err := invokeWorkflowHostCallback(t, hostServices, &proto.InvokeWorkflowOperationRequest{
-		Target:       protoWorkflowPluginTarget("roadmap", "sync"),
-		ExecutionRef: executionRef,
-	})
-	if err != nil {
-		t.Fatalf("invoke workflow host callback: %v", err)
-	}
-	if resp.GetStatus() != http.StatusAccepted || resp.GetBody() != `{"ok":true}` {
-		t.Fatalf("workflow host callback response = %#v", resp)
-	}
-	if got, _ := requestPath.Load().(string); got != "/sync" {
-		t.Fatalf("request path = %q, want %q", got, "/sync")
-	}
-}
-
 func TestValidateManagedWorkflowStartupCallbackUsesPreparedProviderStub(t *testing.T) {
 	t.Parallel()
 
@@ -6442,9 +4792,6 @@ func TestValidateManagedWorkflowStartupCallbackUsesPreparedProviderStub(t *testi
 			cfg.Providers.Workflow = map[string]*config.ProviderEntry{
 				"temporal": {Source: config.ProviderSource{Path: "stub"}},
 			}
-			setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-				Provider: "temporal",
-			})
 
 			factories := validFactories()
 			factories.Workflow = func(_ context.Context, name string, _ yaml.Node, hostServices []providerhost.HostService, deps bootstrap.Deps) (coreworkflow.Provider, error) {
@@ -6527,9 +4874,6 @@ func TestValidateManagedWorkflowStartupInvokesMCPPassthroughPreparedProviders(t 
 	cfg.Providers.Workflow = map[string]*config.ProviderEntry{
 		"temporal": {Source: config.ProviderSource{Path: "stub"}},
 	}
-	setWorkflowFixture(cfg, "roadmap", &workflowFixture{
-		Provider: "temporal",
-	})
 
 	factories := validFactories()
 	factories.Workflow = func(_ context.Context, name string, _ yaml.Node, _ []providerhost.HostService, deps bootstrap.Deps) (coreworkflow.Provider, error) {
