+++ gestaltd/internal/config/config_validate.go
@@ -12,9 +12,7 @@ import (
 	"sort"
 	"strconv"
 	"strings"
-	"time"
 
-	cronv3 "github.com/robfig/cron/v3"
 	"github.com/valon-technologies/gestalt/server/internal/providerenv"
 	"github.com/valon-technologies/gestalt/server/internal/providerpkg"
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
@@ -146,7 +144,7 @@ func ValidateCanonicalStructure(cfg *Config) error {
 			return err
 		}
 	}
-	return validateWorkflowsConfig(cfg)
+	return nil
 }
 
 func validateProviderDevConfig(cfg *Config) error {
@@ -1136,227 +1134,6 @@ func validatePluginS3Bindings(cfg *Config, name string, entry *ProviderEntry) er
 		seen[binding] = struct{}{}
 		envNames[envName] = binding
 		entry.S3[i] = binding
-	}
-	return nil
-}
-
-var workflowScheduleCronParser = cronv3.NewParser(
-	cronv3.Minute | cronv3.Hour | cronv3.Dom | cronv3.Month | cronv3.Dow,
-)
-
-func validateWorkflowScheduleCron(scheduleKey, spec string) error {
-	if _, err := workflowScheduleCronParser.Parse(spec); err != nil {
-		return fmt.Errorf("config validation: workflows.schedules.%s.cron %q is invalid: %w", scheduleKey, spec, err)
-	}
-	return nil
-}
-
-func validateWorkflowsConfig(cfg *Config) error {
-	if len(cfg.Workflows.Schedules) > 0 {
-		normalized := make(map[string]WorkflowScheduleConfig, len(cfg.Workflows.Schedules))
-		for key := range cfg.Workflows.Schedules {
-			schedule := cfg.Workflows.Schedules[key]
-			key = strings.TrimSpace(key)
-			if key == "" {
-				return fmt.Errorf("config validation: workflows.schedules keys must not be empty")
-			}
-			if _, exists := normalized[key]; exists {
-				return fmt.Errorf("config validation: workflows.schedules duplicates %q", key)
-			}
-			if err := validateWorkflowScheduleTarget(cfg, key, &schedule); err != nil {
-				return err
-			}
-			schedule.Provider = strings.TrimSpace(schedule.Provider)
-			providerName, _, err := cfg.EffectiveWorkflowProvider(schedule.Provider)
-			if err != nil {
-				return fmt.Errorf("config validation: workflows.schedules.%s.provider: %w", key, err)
-			}
-			if providerName == "" {
-				return fmt.Errorf("config validation: workflows.schedules.%s.provider is required", key)
-			}
-			schedule.Cron = strings.TrimSpace(schedule.Cron)
-			if schedule.Cron == "" {
-				return fmt.Errorf("config validation: workflows.schedules.%s.cron is required", key)
-			}
-			if err := validateWorkflowScheduleCron(key, schedule.Cron); err != nil {
-				return err
-			}
-			schedule.Timezone = strings.TrimSpace(schedule.Timezone)
-			if schedule.Timezone == "" {
-				schedule.Timezone = "UTC"
-			}
-			if _, err := time.LoadLocation(schedule.Timezone); err != nil {
-				return fmt.Errorf("config validation: workflows.schedules.%s.timezone %q is invalid: %w", key, schedule.Timezone, err)
-			}
-			normalized[key] = schedule
-		}
-		cfg.Workflows.Schedules = normalized
-	}
-
-	if len(cfg.Workflows.EventTriggers) > 0 {
-		normalized := make(map[string]WorkflowEventTriggerConfig, len(cfg.Workflows.EventTriggers))
-		for key := range cfg.Workflows.EventTriggers {
-			trigger := cfg.Workflows.EventTriggers[key]
-			key = strings.TrimSpace(key)
-			if key == "" {
-				return fmt.Errorf("config validation: workflows.eventTriggers keys must not be empty")
-			}
-			if _, exists := normalized[key]; exists {
-				return fmt.Errorf("config validation: workflows.eventTriggers duplicates %q", key)
-			}
-			if err := validateWorkflowEventTriggerTarget(cfg, key, &trigger); err != nil {
-				return err
-			}
-			trigger.Provider = strings.TrimSpace(trigger.Provider)
-			providerName, _, err := cfg.EffectiveWorkflowProvider(trigger.Provider)
-			if err != nil {
-				return fmt.Errorf("config validation: workflows.eventTriggers.%s.provider: %w", key, err)
-			}
-			if providerName == "" {
-				return fmt.Errorf("config validation: workflows.eventTriggers.%s.provider is required", key)
-			}
-			trigger.Match.Type = strings.TrimSpace(trigger.Match.Type)
-			if trigger.Match.Type == "" {
-				return fmt.Errorf("config validation: workflows.eventTriggers.%s.match.type is required", key)
-			}
-			trigger.Match.Source = strings.TrimSpace(trigger.Match.Source)
-			trigger.Match.Subject = strings.TrimSpace(trigger.Match.Subject)
-			normalized[key] = trigger
-		}
-		cfg.Workflows.EventTriggers = normalized
-	}
-	return nil
-}
-
-func validateWorkflowScheduleTarget(cfg *Config, key string, schedule *WorkflowScheduleConfig) error {
-	if schedule == nil {
-		return fmt.Errorf("config validation: workflows.schedules.%s is required", key)
-	}
-	targetPath := "workflows.schedules." + key + ".target"
-	if err := normalizeWorkflowTarget(targetPath, schedule.Target); err != nil {
-		return err
-	}
-	if schedule.Target.Plugin != nil {
-		if _, ok := cfg.Plugins[schedule.Target.Plugin.Name]; !ok {
-			return fmt.Errorf("config validation: %s.plugin.name references unknown plugin %q", targetPath, schedule.Target.Plugin.Name)
-		}
-		return nil
-	}
-	return validateWorkflowAgentConfig(cfg, targetPath+".agent", schedule.Target.Agent)
-}
-
-func validateWorkflowEventTriggerTarget(cfg *Config, key string, trigger *WorkflowEventTriggerConfig) error {
-	if trigger == nil {
-		return fmt.Errorf("config validation: workflows.eventTriggers.%s is required", key)
-	}
-	targetPath := "workflows.eventTriggers." + key + ".target"
-	if err := normalizeWorkflowTarget(targetPath, trigger.Target); err != nil {
-		return err
-	}
-	if trigger.Target.Plugin != nil {
-		if _, ok := cfg.Plugins[trigger.Target.Plugin.Name]; !ok {
-			return fmt.Errorf("config validation: %s.plugin.name references unknown plugin %q", targetPath, trigger.Target.Plugin.Name)
-		}
-		return nil
-	}
-	return validateWorkflowAgentConfig(cfg, targetPath+".agent", trigger.Target.Agent)
-}
-
-func normalizeWorkflowTarget(path string, target *WorkflowTargetConfig) error {
-	if target == nil {
-		return fmt.Errorf("config validation: %s must set exactly one of plugin or agent", path)
-	}
-	hasPlugin := target.Plugin != nil
-	hasAgent := target.Agent != nil
-	if hasPlugin == hasAgent {
-		return fmt.Errorf("config validation: %s must set exactly one of plugin or agent", path)
-	}
-	if hasPlugin {
-		plugin := *target.Plugin
-		plugin.Name = strings.TrimSpace(plugin.Name)
-		plugin.Operation = strings.TrimSpace(plugin.Operation)
-		plugin.Connection = strings.TrimSpace(plugin.Connection)
-		plugin.Instance = strings.TrimSpace(plugin.Instance)
-		if plugin.Name == "" {
-			return fmt.Errorf("config validation: %s.plugin.name is required", path)
-		}
-		if plugin.Operation == "" {
-			return fmt.Errorf("config validation: %s.plugin.operation is required", path)
-		}
-		target.Plugin = &plugin
-		return nil
-	}
-	return nil
-}
-
-func validateWorkflowAgentConfig(cfg *Config, path string, agent *WorkflowAgentConfig) error {
-	if agent == nil {
-		return fmt.Errorf("config validation: %s is required", path)
-	}
-	agent.Provider = strings.TrimSpace(agent.Provider)
-	providerName, _, err := cfg.EffectiveAgentProvider(agent.Provider)
-	if err != nil {
-		return fmt.Errorf("config validation: %s.provider: %w", path, err)
-	}
-	if providerName == "" {
-		return fmt.Errorf("config validation: %s.provider is required", path)
-	}
-	agent.Provider = providerName
-	agent.Model = strings.TrimSpace(agent.Model)
-	agent.Prompt = strings.TrimSpace(agent.Prompt)
-	for i := range agent.Messages {
-		agent.Messages[i].Role = strings.TrimSpace(agent.Messages[i].Role)
-		agent.Messages[i].Text = strings.TrimSpace(agent.Messages[i].Text)
-	}
-	if agent.Prompt == "" && len(agent.Messages) == 0 {
-		return fmt.Errorf("config validation: %s.prompt or messages is required", path)
-	}
-	hasSystemTool := false
-	for i := range agent.Tools {
-		if strings.TrimSpace(agent.Tools[i].System) != "" {
-			hasSystemTool = true
-			break
-		}
-	}
-	for i := range agent.Tools {
-		tool := &agent.Tools[i]
-		tool.System = strings.TrimSpace(tool.System)
-		tool.Plugin = strings.TrimSpace(tool.Plugin)
-		if tool.System == "" && tool.Plugin == "" {
-			return fmt.Errorf("config validation: %s.tools[%d].plugin or system is required", path, i)
-		}
-		if tool.System != "" && tool.Plugin != "" {
-			return fmt.Errorf("config validation: %s.tools[%d] must set exactly one of plugin or system", path, i)
-		}
-		tool.Operation = strings.TrimSpace(tool.Operation)
-		tool.Connection = strings.TrimSpace(tool.Connection)
-		tool.Instance = strings.TrimSpace(tool.Instance)
-		tool.Title = strings.TrimSpace(tool.Title)
-		tool.Description = strings.TrimSpace(tool.Description)
-		if tool.System != "" {
-			if tool.System != "workflow" {
-				return fmt.Errorf("config validation: %s.tools[%d].system references unknown system %q", path, i, tool.System)
-			}
-			if tool.Operation == "" {
-				return fmt.Errorf("config validation: %s.tools[%d].operation is required for system tool refs", path, i)
-			}
-			if tool.Connection != "" || tool.Instance != "" {
-				return fmt.Errorf("config validation: %s.tools[%d] system refs cannot include connection or instance", path, i)
-			}
-			continue
-		}
-		if _, ok := cfg.Plugins[tool.Plugin]; !ok {
-			return fmt.Errorf("config validation: %s.tools[%d].plugin references unknown plugin %q", path, i, tool.Plugin)
-		}
-		if hasSystemTool && tool.Operation == "" {
-			return fmt.Errorf("config validation: %s.tools[%d].operation is required when workflow system tools are delegated", path, i)
-		}
-	}
-	agent.Timeout = strings.TrimSpace(agent.Timeout)
-	if agent.Timeout != "" {
-		if _, err := time.ParseDuration(agent.Timeout); err != nil {
-			return fmt.Errorf("config validation: %s.timeout %q is invalid: %w", path, agent.Timeout, err)
-		}
 	}
 	return nil
 }
